+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Ipfs web uploader
+- Ipfs editor uploader
 
 ## [1.5.0] - 2023-06-22
 ### Fixed